Match streaming test files to their source files
Run lives in coordinator.go, so its tests move from run_test.go to
coordinator_test.go (tagged, like coordinator.go); the StreamOptions test
moves to a new options_test.go, keeping it untagged like options.go so it
still runs in untagged builds.
