+++ services/motion/builtin/streaming/options_test.go
@@ -1,0 +1,40 @@
+// Package streaming implements the ability to stream joint positions to an arm resource.
+package streaming
+
+import (
+	"testing"
+
+	"go.viam.com/test"
+)
+
+func TestStreamOptionsDefaultsAndValidate(t *testing.T) {
+	valid := NewDefaultOptions()
+	test.That(t, valid.TargetRunwayInArmMs, test.ShouldEqual, defaultTargetRunwayInArmMs)
+	test.That(t, valid.SendToArmIntervalMs, test.ShouldEqual, defaultSendToArmIntervalMs)
+	test.That(t, valid.VelLimitDegPerSec, test.ShouldEqual, defaultVelLimitDegPerSec)
+	test.That(t, valid.AccelLimitDegPerSec2, test.ShouldEqual, defaultAccelLimitDegPerSec2)
+	test.That(t, valid.Validate(), test.ShouldBeNil)
+
+	// The zero value does not validate.
+	test.That(t, (&StreamOptions{}).Validate(), test.ShouldNotBeNil)
+
+	// Each Validate rule, violated in isolation against the otherwise-valid base.
+	for _, tc := range []struct {
+		name   string
+		mutate func(*StreamOptions)
+	}{
+		{"zero runway", func(o *StreamOptions) { o.TargetRunwayInArmMs = 0 }},
+		{"negative runway", func(o *StreamOptions) { o.TargetRunwayInArmMs = -1 }},
+		// A zero send interval is invalid (division by zero when converting to Hz).
+		{"zero send interval", func(o *StreamOptions) { o.SendToArmIntervalMs = 0 }},
+		{"interval not less than runway", func(o *StreamOptions) { o.SendToArmIntervalMs = o.TargetRunwayInArmMs }},
+		{"zero vel limit", func(o *StreamOptions) { o.VelLimitDegPerSec = 0 }},
+		{"negative vel limit", func(o *StreamOptions) { o.VelLimitDegPerSec = -1 }},
+		{"zero accel limit", func(o *StreamOptions) { o.AccelLimitDegPerSec2 = 0 }},
+		{"negative accel limit", func(o *StreamOptions) { o.AccelLimitDegPerSec2 = -1 }},
+	} {
+		bad := valid
+		tc.mutate(&bad)
+		test.That(t, bad.Validate(), test.ShouldNotBeNil)
+	}
+}
